+++ cmd/operator-sdk/add/controller.go
@@ -25,7 +25,7 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var k8sAPIImport string
+var customAPIImport string
 
 func newAddControllerCmd() *cobra.Command {
 	controllerCmd := &cobra.Command{
@@ -59,7 +59,7 @@ Example:
 	if err := controllerCmd.MarkFlagRequired("kind"); err != nil {
 		log.Fatalf("Failed to mark `kind` flag for `add controller` subcommand as required")
 	}
-	controllerCmd.Flags().StringVar(&k8sAPIImport, "k8s-api-import", "", `External Kubernetes resource import path of the form "host.com/repo/path[=import_identifier]". import_identifier is optional`)
+	controllerCmd.Flags().StringVar(&customAPIImport, "custom-api-import", "", `External Kubernetes resource import path of the form "host.com/repo/path[=import_identifier]". import_identifier is optional`)
 
 	return controllerCmd
 }
@@ -87,7 +87,7 @@ func controllerRun(cmd *cobra.Command, args []string) error {
 	s := &scaffold.Scaffold{}
 
 	err = s.Execute(cfg,
-		&scaffold.ControllerKind{Resource: r, K8sImport: k8sAPIImport},
+		&scaffold.ControllerKind{Resource: r, CustomImport: customAPIImport},
 		&scaffold.AddController{Resource: r},
 	)
 	if err != nil {